Cambios menores en role

+++ BurguerAvenida/src/main/java/com/pd/dao/security/RoleDao.java
@@ -7,6 +7,6 @@ import com.pd.model.security.Role;
 import com.pd.model.security.RoleName;
 
 @Repository
-public interface RoleDao extends CrudRepository<Role, Long> {
+public interface RoleDao extends CrudRepository<Role, Integer> {
 	Role findByName(RoleName name);
 }
